refactor: extract improvement parser
